requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
WTForms==3.1.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bleach==6.1.0
eventlet==0.33.3
redis==5.0.1
pandas==2.1.3
openpyxl==3.1.2